+++ src/types/random.rs
@@ -25,7 +25,7 @@ impl fmt::Debug for JanetRng {
     }
 }
 
-impl IsJanetAbstract for JanetRng {
+unsafe impl IsJanetAbstract for JanetRng {
     type Get = Self;
 
     const SIZE: usize = mem::size_of::<Self>();